LPD-82337 Move tests to a single file
