+++ modules/test/playwright/tests/object-web/forms-integration/config.ts
@@ -1,0 +1,12 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export const config = {
+	name: 'object-web.forms-integration',
+	testDir: 'tests/object-web/forms-integration',
+	use: {
+		testIdAttribute: 'data-qa-id',
+	},
+};
