+++ modules/test/playwright/tests/object-web/forms-integration/test.properties
@@ -1,0 +1,1 @@
+testray.main.component.name=Object
